Determine actual difficulty calculation coefficients
After about an hour of messing around with strain decay bases and skill
multipliers, I settled with the these values. They're not perfect, as
fast maps with bursts (and not streams) still tend to be somewhat
underrated compared to slow maps with a couple streams (i.e. Skyclad
no Kanokusha).